+++ app/api.js
@@ -1,5 +1,5 @@
 const wreck = require('@hapi/wreck')
-const config = require('../config')
+const config = require('./config')
 
 async function get (url, token) {
   return wreck.get(`${config.paymentsEndpoint}${url}`, getConfiguration(token))
