+++ app/insights.js
