+++ app/payment-holds.js
@@ -1,4 +1,4 @@
-const api = require('../api')
+const api = require('./api')
 
 const getResponse = async (url) => {
   try {
