+++ app/plugins/logging.js
@@ -5,6 +5,6 @@ module.exports = {
   options: {
     logPayload: true,
     prettyPrint: config.isDev,
-    level: config.isDev ? 'debug' : 'warn'
+    level: 'warn'
   }
 }
