+++ app/server.js
@@ -1,8 +1,5 @@
 const hapi = require('@hapi/hapi')
 const config = require('./config')
-const catbox = config.useRedis ? require('@hapi/catbox-redis') : require('@hapi/catbox-memory')
-const catboxOptions = config.useRedis ? config.cacheConfig.redisCatboxOptions : {}
-const cache = require('./cache')
 
 async function createServer () {
   // Create the hapi server
@@ -17,14 +14,7 @@ async function createServer () {
     },
     router: {
       stripTrailingSlash: true
-    },
-    cache: [{
-      name: 'session',
-      provider: {
-        constructor: catbox,
-        options: catboxOptions
-      }
-    }]
+    }
   })
 
   // Register the plugins
@@ -32,15 +22,12 @@ async function createServer () {
   await server.register(require('./plugins/views'))
   await server.register(require('./plugins/router'))
   await server.register(require('./plugins/error-pages'))
-  // await server.register(require('./plugins/session'))
   await server.register(require('./plugins/crumb'))
 
   if (config.isDev) {
     await server.register(require('blipp'))
     await server.register(require('./plugins/logging'))
   }
-
-  cache.setup(server)
 
   return server
 }
