--- app/views/tag-manager/head-content.njk
@@ -1,5 +1,0 @@
-{% if googleTagManagerKey %}
-<!-- Google Tag Manager (noscript) -->
-<noscript><iframe src="https://www.googletagmanager.com/ns.html?id={{ googleTagManagerKey }}" height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
-<!-- End Google Tag Manager (noscript) -->
-{% endif %}
